remember to close and unmap buffer after drawing
So when you open wev and wiggle it about a bit in an interactive
resize it eats a bunch of memory before dieing of too many open files.
It looks like it was just forgotten to clean up the wl_buffer when
we're done. I think this is all that's needed.